更新数据

+++ src/modules/mavlink/mavlink_timesync.cpp
@@ -41,6 +41,8 @@
 #include "mavlink_timesync.h"
 #include "mavlink_main.h"
 
+#include <stdlib.h>
+
 MavlinkTimesync::MavlinkTimesync(Mavlink *mavlink) :
 	_mavlink(mavlink)
 {
@@ -107,10 +109,10 @@ MavlinkTimesync::handle_message(const mavlink_message_t *msg)
 						// Filter gain scheduling
 						if (!sync_converged()) {
 							// Interpolate with a sigmoid function
-							float progress = ((float)_sequence) / CONVERGENCE_WINDOW;
-							float p = 1.0f - expf(0.5f * (1.0f - 1.0f / (1.0f - progress)));
-							_filter_alpha = p * (float)ALPHA_GAIN_FINAL + (1.0f - p) * (float)ALPHA_GAIN_INITIAL;
-							_filter_beta = p * (float)BETA_GAIN_FINAL + (1.0f - p) * (float)BETA_GAIN_INITIAL;
+							double progress = (double)_sequence / (double)CONVERGENCE_WINDOW;
+							double p = 1.0 - exp(0.5 * (1.0 - 1.0 / (1.0 - progress)));
+							_filter_alpha = p * ALPHA_GAIN_FINAL + (1.0 - p) * ALPHA_GAIN_INITIAL;
+							_filter_beta = p * BETA_GAIN_FINAL + (1.0 - p) * BETA_GAIN_INITIAL;
 
 						} else {
 							_filter_alpha = ALPHA_GAIN_FINAL;
